Add files via upload
To run this web application, first you need to run the apache server on your linux.  You can take a reference from the text file and the other python script is the main script for the web application. You need to save this python script, pyth.py, in the cgi-bin folder to run this web application.